requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
pyttsx3==2.90
SpeechRecognition==3.10.0
python-jose==3.3.0
jinja2==3.1.2 